% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputValideringsfunktioner.R
\encoding{UTF-8}
\name{inputvalManglerOgVaerdier}
\alias{inputvalManglerOgVaerdier}
\title{Inputvalidering (Mangler og værdier)}
\usage{
inputvalManglerOgVaerdier(
  datasaet,
  datasaetTilladtMangel,
  datasaetVaerdiListe,
  sti,
  udskrivTekstTilRapport = TRUE,
  datasaetNavn = NA,
  meldFejlVedMangler = FALSE,
  meldFejlVedIkkeTilladteVaerdier = FALSE
)
}
\arguments{
\item{datasaet}{en data.frame, Datasættet som skal gennemses for mangler og værdier}

\item{datasaetTilladtMangel}{En data.frame med to kolonner, hvor første kolonne svarer til variabelnavnene og anden kolonne svarer til den maksimale andel der er tilladt at mangle i den enkelte variabel.
Rækkefølgen af kolonnerne har betydning for om funktionen virker.}

\item{datasaetVaerdiListe}{En data.frame med to kolonner, hvor første kolonne svarer til variabelnavnene og anden kolonne svarer til navnet på den liste over tilladte/acceptable værdier for variablen.
Rækkefølgen af kolonnerne har betydning for om funktionen virker.}

\item{sti}{en chracter string, Stien til den rapport, som outputtet af funktionen skal sendes til. Rapporten skal være en txt-fil hvis udskrivTekstTilRapport er TRUE ellers skal det være stien til mappen som tabellerne skal udskrives til.}

\item{udskrivTekstTilRapport}{TRUE/FALSE værdi, Hvis TRUE så udskrives teksterne til txt-filen som er angivet i "sti". HVis FALSE bør sti være en mappesti.}

\item{datasaetNavn}{en chracter string, Brugerdefineret navn til at definere datasættet i rapporten. Hvis NA (default) bruges objektnavnet på datasættet fra miljøet.}

\item{meldFejlVedMangler}{TRUE/FALSE værdi, Hvis TRUE melder funktionen fejl, hvis der mangler mere end tilladt i mindst en variabel i datasættet. Fejlmeldingen sker efter tabellen er udskrevet.}

\item{meldFejlVedIkkeTilladteVaerdier}{TRUE/FALSE værdi, Hvis TRUE melder funktionen fejl, hvis der findes ikke tilladte værdier i mindst en variabel i datasættet. Fejlmeldingen sker efter tabellen er udskrevet.}
}
\value{
To sæt af pdf filer, hvor den ene viser en liste over variable og hvor meget der mangler i hver variabel og den anden viser de værdier der ikke findes i værdilisten for variabelen (hvis der findes sådanne værdier).
}
\description{
Funktion til at gennemse datasættet for mangler og værdisæt for hver variabel med udgangspunkt i den tilhørende formatfil.
    Funktionen består af to dele:\cr
    \itemize{
    \item Første del gennemser datasættet for mangler og udskriver en tabel med andelen af mangler og bemærkninger.\cr
    Bemærkninger sættes kun på de variable, hvor der mangler mere end tilladt.\cr
    Funktionen melder herefter fejl i tilfælde af, at der mangler mere end tilladt i mindst en variabel, hvis dette ønskes.
    \item Anden del tester værdierne i datasættets variable mod den tilhørende værdiliste (defineret i formatfilen).\cr
    Der udskrives en tabel for hver enkelt variabel med de værdier som ikke er tilladt/accepteret, hvis der findes sådanne værdier.\cr
    Funktionen melder herefter fejl i tilfælde af at der findes mindst en variabel med ikke tilladte værdier, hvis dette ønskes.
    }
}
\details{
Funktionen udskriver tekst til rapporten i stien, hvis udskrivTekstTilRapport er valgt.
    De to pdf filer, der genereres af funktionen, bliver gemt under samme manppe som rapporten eller mappen som er angivet i "sti", hvis udskrivTekstTilRapport ikke er valgt.
}
